Enclose packet output in debugging statements.
Also set DEBUGGING to False in CMake so that the server will by
default not include the extra output while running (which will be
useful for those who want to compile it straight to production).

I chose to keep the statements in there for later in case there
are problems or things need to be added since they were carefully
placed and should be representative of what's going on.