Add Tethealla's login server as-is to project.
Also create apropriate targets in CMake so that the files are added
to the project when generated. At this point it's nearly guaranteed
that the server source won't work on *nix (though it compiles fine
on Windows) so next steps are to start replacing all Windows-specific
includes/functions with nix equivalents.