Refactor project-wide CMake build rules.
Moved common CMake modules out to the root-level directory so that
all targets can access them. Added a find module for Jansson and wrote
one for MySQLConnectorC as well as correcting the linking for Jansson.
As of now, the project will build and run successfully (for all targets)
via Make.