General improvements.
Worst commit message ever, I really shouldn't do this. But this change
got out of hand and now includes a bunch of related changes that
are difficult to pick apart and I haven't looked at this in a month.

The purpose of this commit is to be able to create another branch
to attempt a different approach at refactoring and see how far it
gets. Note that at this point the project should compile and run,
but gets caught up somewhere in the client connection of the login
server (like I said, it's been a while).