Fix and refactor lots of things.
I'm really getting bad about these messages. But really, lots of
misc fixes with emphasis on implementing the handling of the 0x93
login packet more effectively. The security packet (E6) still needs
to be fixed for when an account login is successful as it currently
causes the client to hang and refuse to disconnect. Removing it works
but messes up later 0x93 packets. Redirect (0x19) works as expected
if the security packet is removed.