Remove big endian support and onstant size defs.
The BE wasn't used consistently and I think it cluttered up the code
unnecessarily since I think it's more likely that the server will be
run on LE machines. The constant size definitions also cluttered up
the header file for such a small optimization, so they've been dropped
in favor of one call to sizeof each.